+++ src/test/java/io/github/classgraph/features/ClassTypeAnnotation.java
@@ -46,6 +46,14 @@ class ClassTypeAnnotation {
     }
 
     /***/
+    private interface C {
+    }
+
+    /***/
+    private interface BSubC extends C {
+    }
+
+    /***/
     private static class E<T> extends @P Z implements @Q A, @R B {
     }
 
@@ -63,6 +71,10 @@ class ClassTypeAnnotation {
 
     /***/
     private static class I implements @Q B, @R A {
+    }
+
+    /***/
+    private static class J implements @Q BSubC, @R A {
     }
 
     @Test
@@ -108,6 +120,15 @@ class ClassTypeAnnotation {
                     .isEqualTo("private static class " + I.class.getName() + " implements "
                             + ClassTypeAnnotation.class.getName() + "$@" + Q.class.getName() + " "
                             + B.class.getSimpleName() + ", " + ClassTypeAnnotation.class.getName() + "$@"
+                            + R.class.getName() + " " + A.class.getSimpleName());
+
+            // The synthesized type descriptor must list the directly implemented
+            // interfaces, in classfile order, since the class type annotation targets
+            // index into the classfile's own interfaces[] array
+            assertThat(scanResult.getClassInfo(J.class.getName()).getTypeSignatureOrTypeDescriptor().toString())
+                    .isEqualTo("private static class " + J.class.getName() + " implements "
+                            + ClassTypeAnnotation.class.getName() + "$@" + Q.class.getName() + " "
+                            + BSubC.class.getSimpleName() + ", " + ClassTypeAnnotation.class.getName() + "$@"
                             + R.class.getName() + " " + A.class.getSimpleName());
         }
     }